requirements.txt
django>=4.2
djangorestframework
requests
psycopg2-binary
